# internal/modules/roles/handler/role_handler.go
package handler

import (
	"log"
	"strconv"

	"github.com/gofiber/fiber/v2"
	roleModel "github.com/jumayevgadam/tsu-toleg/internal/models/role"
	"github.com/jumayevgadam/tsu-toleg/pkg/errlst"
	"github.com/jumayevgadam/tsu-toleg/pkg/errlst/tracing"
	"github.com/jumayevgadam/tsu-toleg/pkg/reqvalidator"
	"go.opentelemetry.io/otel"
	"go.opentelemetry.io/otel/codes"
)

// AddRole handleris method adds a new role to the system and returns the created role's id.
// @Summary Add-Role.
// @Description creates a new role and returns its id.
// @Tags Roles
// @ID add-role
// @Accept multipart/form-data
// @Produce json
// @Param roleReq formData roleModel.DTO true "role request payload"
// @Success 200 {integer} integer 1
// @Failure 400 {object} errlst.RestErr
// @Failure 500 {object} errlst.RestErr
// @Router /role/create [post]
func (h *RoleHandler) AddRole() fiber.Handler {
	return func(c *fiber.Ctx) error {
		ctx, span := otel.Tracer("[RoleHandler][AddRole]").Start(c.Context(), "[RoleHandler]")
		defer span.End()
		var roleReq roleModel.DTO

		if err := reqvalidator.ReadRequest(c, &roleReq); err != nil {
			tracing.EventErrorTracer(span, err, "[AddRole][Handler]")
			return errlst.Response(c, err)
		}

		roleID, err := h.service.AddRole(ctx, roleReq)
		if err != nil {
			tracing.EventErrorTracer(span, err, "[AddRole][Handler]")
			return errlst.Response(c, err)
		}

		return c.Status(fiber.StatusCreated).JSON(fiber.Map{
			"role":   roleReq.RoleName,
			"roleID": roleID,
		})
	}
}

// GetRole handler method fetches a role by its id and returns its details.
// @Summary Get-Role
// @Description retrieve role by identified id.
// @Tags Roles
// @ID get-role
// @Accept multipart/form-data
// @Produce json
// @Param id path int true "id"
// @Success 200 {object} roleModel.DTO
// @Failure 400 {object} errlst.RestErr
// @Failure 500 {object} errlst.RestErr
// @Router /role/{id} [get]
func (h *RoleHandler) GetRole() fiber.Handler {
	return func(c *fiber.Ctx) error {
		ctx, span := otel.Tracer("[RoleHandler]").Start(c.Context(), "[GetRole]")
		defer span.End()

		roleID, err := strconv.Atoi(c.Params("id"))
		if err != nil {
			tracing.EventErrorTracer(span, err, "getting error in roleID taking")
			return errlst.Response(c, err)
		}

		role, err := h.service.GetRole(ctx, roleID)
		if err != nil {
			tracing.EventErrorTracer(span, err, "service_error")
			return errlst.Response(c, err)
		}

		span.SetStatus(codes.Ok, "successfully got role by ID")
		return c.Status(fiber.StatusOK).JSON(role)
	}
}

// GetRoles handler method fetches and returns a list of all roles.
// @Summary List-Roles
// @Description list roles
// @Tags Roles
// @ID list-roles
// @Produce json
// @Success 200 {object} []roleModel.DTO
// @Failure 400 {object} errlst.RestErr
// @Failure 500 {object} errlst.RestErr
// @Router /role/get-all [get]
func (h *RoleHandler) GetRoles() fiber.Handler {
	return func(c *fiber.Ctx) error {
		ctx, span := otel.Tracer("[RoleHandler]").Start(c.Context(), "GetRoles")
		defer span.End()

		roles, err := h.service.GetRoles(ctx)
		if err != nil {
			log.Println(err)
			tracing.EventErrorTracer(span, err, errlst.ErrInternalServer.Error())
			return errlst.Response(c, err)
		}

		span.SetStatus(codes.Ok, "successfully got roles")
		return c.Status(200).JSON(roles)
	}
}

// DeleteRole handler method deletes a role from the system identified by the given id.
// @Summary Delete-Role
// @Description deletes role using identified role id.
// @Tags Roles
// @ID delete-role
// @Accept multipart/form-data
// @Produce json
// @Param id path int true "id"
// @Success 200 {string} string "role successfully deleted"
// @Failure 400 {object} errlst.RestErr
// @Failure 500 {object} errlst.RestErr
// @Router /role/{id} [delete]
func (h *RoleHandler) DeleteRole() fiber.Handler {
	return func(c *fiber.Ctx) error {
		ctx, span := otel.Tracer("[RoleHandler]").Start(c.Context(), "[DeleteRole]")
		defer span.End()

		roleID, err := strconv.Atoi(c.Params("id"))
		if err != nil {
			tracing.EventErrorTracer(span, err, errlst.ErrBadQueryParams.Error())
			return errlst.Response(c, err)
		}

		if err := h.service.DeleteRole(ctx, roleID); err != nil {
			tracing.EventErrorTracer(span, err, errlst.ErrInternalServer.Error())
			return errlst.Response(c, err)
		}

		span.SetStatus(codes.Ok, "successfully deleted role")
		return c.Status(fiber.StatusNoContent).JSON("Successfully deleted role")
	}
}

// UpdateRole handler method updates an existing role based on the provided id and new role data.
// @Summary Update-Role
// @Description updating roles only giving rolename and do not use roleID, only can change roleName by given id.
// @Tags Roles
// @ID update-role
// @Accept multipart/form-data
// @Produce json
// @Param id path int true "id"
// @Param roleReq formData roleModel.DTO true "update request for roles"
// @Success 200 {string} string "role successfully updated"
// @Failure 400 {object} errlst.RestErr
// @Failure 500 {object} errlst.RestErr
// @Router /role/{id} [put]
func (h *RoleHandler) UpdateRole() fiber.Handler {
	return func(c *fiber.Ctx) error {
		ctx, span := otel.Tracer("[RoleHandler]").Start(c.Context(), "[UpdateRole]")
		defer span.End()

		roleID, err := strconv.Atoi(c.Params("id"))
		if err != nil {
			tracing.EventErrorTracer(span, err, errlst.ErrBadRequest.Error())
			return errlst.Response(c, err)
		}

		var roleReq roleModel.DTO
		if err := reqvalidator.ReadRequest(c, &roleReq); err != nil {
			if roleReq.RoleName == "" {
				return c.Status(fiber.StatusOK).JSON("update structure has no value")
			}
			tracing.EventErrorTracer(span, err, errlst.ErrFieldValidation.Error())

			return errlst.Response(c, err)
		}
		roleReq.ID = roleID

		res, err := h.service.UpdateRole(ctx, roleReq)
		if err != nil {
			tracing.EventErrorTracer(span, err, errlst.ErrInternalServer.Error())
			return errlst.Response(c, err)
		}

		span.SetStatus(codes.Ok, "Successfully updated role")
		return c.Status(fiber.StatusOK).JSON(res)
	}
}